Move methods for planetgen from planet impl
